requirements.txt
Flask==2.0.2
boto3
pytz
python-dotenv